Add data-url to api-visualelement to make frontend-code not fail.
