+++ concept-api/src/main/scala/no/ndla/conceptapi/service/ReadService.scala
@@ -7,19 +7,11 @@
 
 package no.ndla.conceptapi.service
 
-import io.lemonlabs.uri.{Path, Url}
-import no.ndla.conceptapi.ConceptApiProperties.externalApiUrls
 import no.ndla.conceptapi.model.api
 import no.ndla.conceptapi.model.api.NotFoundException
-import no.ndla.conceptapi.model.domain.Concept
 import no.ndla.conceptapi.repository.{DraftConceptRepository, PublishedConceptRepository}
 import no.ndla.language.Language
-import no.ndla.validation.EmbedTagRules.ResourceHtmlEmbedTag
-import no.ndla.validation.HtmlTagRules.{jsoupDocumentToString, stringToJsoupDocument}
-import no.ndla.validation.{ResourceType, TagAttributes}
-import org.jsoup.nodes.Element
 
-import scala.jdk.CollectionConverters._
 import scala.util.{Failure, Success, Try}
 
 trait ReadService {
@@ -29,7 +21,7 @@ trait ReadService {
   class ReadService {
 
     def conceptWithId(id: Long, language: String, fallback: Boolean): Try[api.Concept] =
-      draftConceptRepository.withId(id).map(addUrlOnVisualElement) match {
+      draftConceptRepository.withId(id).map(converterService.addUrlOnVisualElement) match {
         case Some(concept) =>
           converterService.toApiConcept(concept, language, fallback)
         case None =>
@@ -37,54 +29,12 @@ trait ReadService {
       }
 
     def publishedConceptWithId(id: Long, language: String, fallback: Boolean): Try[api.Concept] =
-      publishedConceptRepository.withId(id).map(addUrlOnVisualElement) match {
+      publishedConceptRepository.withId(id).map(converterService.addUrlOnVisualElement) match {
         case Some(concept) =>
           converterService.toApiConcept(concept, language, fallback)
         case None =>
           Failure(NotFoundException(s"A concept with id $id was not found with language '$language'."))
       }
-
-    def addUrlOnVisualElement(concept: Concept): Concept = {
-      val visualElementWithUrls =
-        concept.visualElement.map(visual => visual.copy(visualElement = addUrlOnElement(visual.visualElement)))
-      concept.copy(visualElement = visualElementWithUrls)
-    }
-
-    private[service] def addUrlOnElement(content: String): String = {
-      val doc = stringToJsoupDocument(content)
-
-      val embedTags = doc.select(s"$ResourceHtmlEmbedTag").asScala.toList
-      embedTags.foreach(addUrlOnEmbedTag)
-      jsoupDocumentToString(doc)
-    }
-
-    private def addUrlOnEmbedTag(embedTag: Element): Unit = {
-      val typeAndPathOption = embedTag.attr(TagAttributes.DataResource.toString) match {
-        case resourceType
-            if resourceType == ResourceType.H5P.toString
-              && embedTag.hasAttr(TagAttributes.DataPath.toString) =>
-          val path = embedTag.attr(TagAttributes.DataPath.toString)
-          Some((resourceType, path))
-
-        case resourceType if embedTag.hasAttr(TagAttributes.DataResource_Id.toString) =>
-          val id = embedTag.attr(TagAttributes.DataResource_Id.toString)
-          Some((resourceType, id))
-        case _ =>
-          None
-      }
-
-      typeAndPathOption match {
-        case Some((resourceType, path)) =>
-          val baseUrl = Url.parse(externalApiUrls(resourceType))
-          val pathParts = Path.parse(path).parts
-
-          embedTag.attr(
-            s"${TagAttributes.DataUrl}",
-            baseUrl.addPathParts(pathParts).toString
-          )
-        case _ =>
-      }
-    }
 
     def allSubjects(draft: Boolean = false): Try[Set[String]] = {
       if (draft) {
