+++ concept-api/src/test/scala/no/ndla/conceptapi/service/search/PublishedConceptSearchServiceTest.scala
@@ -138,7 +138,7 @@ class PublishedConceptSearchServiceTest
     subjectIds = Set("urn:subject:2"),
     visualElement = List(
       VisualElement(
-        """<embed data-resource="image" data-url="test.url" /><embed data-resource="video" data-url="test.url2" data-resource_id="test.id2" />""",
+        """<embed data-resource="image" data-url="test.url" /><embed data-resource="brightcove" data-url="test.url2" data-videoid="test.id2" />""",
         "nb"))
   )
 
@@ -564,7 +564,7 @@ class PublishedConceptSearchServiceTest
   test("that search on embedResource matches visual element") {
     val Success(search) =
       publishedConceptSearchService.all(
-        searchSettings.copy(embedResource = Some("video"), searchLanguage = Language.AllLanguages))
+        searchSettings.copy(embedResource = Some("brightcove"), searchLanguage = Language.AllLanguages))
 
     search.totalCount should be(1)
     search.results.head.id should be(10)
@@ -593,7 +593,7 @@ class PublishedConceptSearchServiceTest
   test("that search on query parameter as embedResource matches visual element") {
     val Success(search) =
       publishedConceptSearchService.matchingQuery(
-        "video",
+        "brightcove",
         searchSettings.copy()
       )
